requirements.txt
tensorflow
fastapi
uvicorn
pydantic
pillow
numpy
requests
streamlit